Move TxToInteropAccessList to Interop Types package
